Swap paymentInformation back to 526
