+++ src/applications/disability-benefits/526EZ/config/form.js
@@ -26,7 +26,7 @@ import treatmentAddressUiSchema from '../pages/treatmentAddress';
 import {
   uiSchema as paymentInfoUiSchema,
   schema as paymentInfoSchema,
-} from '../../all-claims/pages/paymentInformation';
+} from '../pages/paymentInformation';
 
 import {
   uiSchema as reservesNationalGuardUISchema,
