+++ src/applications/disability-benefits/all-claims/constants.js
@@ -96,18 +96,6 @@ export const ADDRESS_TYPES = {
   forwardingAddress: 'forwardingAddress',
 };
 
-export const accountTitleLabels = {
-  CHECKING: 'Checking Account',
-  SAVINGS: 'Savings Account',
-  NOBANK: 'No Bank Account',
-};
-
-export const NOBANK = 'NOBANK';
-
-export const PENDING = 'PENDING';
-export const RESOLVED = 'RESOLVED';
-export const REJECTED = 'REJECTED';
-
 export const HOMELESSNESS_TYPES = {
   atRisk: 'atRisk',
   homeless: 'homeless',
